Add files via upload
Following File is a google collab file for a basic movie recommendation based on content based movie recommendation that takes in a movie that a user likes 
and then analyzes it to get the movies content such as genre, cast, director and keywords etc based on similarity score. Project was made as a simple mini project during 3rd year of my bachelors.